backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.41
aiosqlite==0.21.0
pydantic==2.11.3
pydantic-settings==2.9.1
httpx==0.28.1
PyJWT==2.10.1
bcrypt==4.3.0
tenacity==9.1.2
python-multipart==0.0.20
aiofiles==24.1.0
